+++ src/test/java/posmy/argos/desk/history/domain/DeskOccupiedHistoryRepositoryTests.java
@@ -1,6 +1,5 @@
 package posmy.argos.desk.history.domain;
 
-import posmy.argos.containers.mongodb.MongoDBTestContainerSetup;
 import posmy.argos.desk.domain.DeskLocation;
 
 import java.util.List;
@@ -9,6 +8,11 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static posmy.argos.desk.helper.DeskTestHelper.create;
 
@@ -21,10 +25,20 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author Rashidi Zin
  */
 @DataMongoTest
-class DeskOccupiedHistoryRepositoryTests extends MongoDBTestContainerSetup {
+@Testcontainers
+class DeskOccupiedHistoryRepositoryTests {
+
+    @Container
+    private static final MongoDBContainer container = new MongoDBContainer();
 
     @Autowired
     private DeskOccupiedHistoryRepository repository;
+    
+    @DynamicPropertySource 
+    static void setupMongoDB(DynamicPropertyRegistry registry) {
+        registry.add("spring.data.mongodb.host", () -> container.getContainerIpAddress());
+        registry.add("spring.data.mongodb.port", () -> container.getMappedPort(27017));
+    }
     
     @Test
     void findFirstByLocationOrderByEndDesc() {
